+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/AndroidComponentsConfig.kt
@@ -26,7 +26,6 @@ import io.sentry.android.gradle.snapshot.GenerateSnapshotTestsTask
 import io.sentry.android.gradle.sourcecontext.OutputPaths
 import io.sentry.android.gradle.sourcecontext.SourceContext
 import io.sentry.android.gradle.tasks.GenerateDistributionPropertiesTask
-import io.sentry.android.gradle.tasks.GenerateSentryBuildTimeOptionsTask
 import io.sentry.android.gradle.tasks.InjectSentryMetaPropertiesIntoAssetsTask
 import io.sentry.android.gradle.tasks.PropertiesFileOutputTask
 import io.sentry.android.gradle.tasks.SentryGenerateIntegrationListTask
@@ -36,6 +35,7 @@ import io.sentry.android.gradle.tasks.SentryUploadProguardMappingsTask
 import io.sentry.android.gradle.tasks.SentryUploadSnapshotsTask
 import io.sentry.android.gradle.tasks.configureNativeSymbolsTask
 import io.sentry.android.gradle.tasks.dependencies.SentryExternalDependenciesReportTaskV2
+import io.sentry.android.gradle.tasks.optimization.GenerateSentryBuildTimeOptionsTask
 import io.sentry.android.gradle.telemetry.SentryTelemetryService
 import io.sentry.android.gradle.util.AgpVersions
 import io.sentry.android.gradle.util.GroovyCompat
@@ -209,6 +209,8 @@ fun ApplicationAndroidComponentsExtension.configure(
             project,
             "${variant.name}RuntimeClasspath",
             variant.name.capitalized,
+            variant.artifacts.get(SingleArtifact.MERGED_MANIFEST),
+            sentryTelemetryProvider,
           )
         val javaSources = variant.sources.java
         if (buildTimeOptionsTask != null && javaSources != null) {